refactor(providers): simplify Bedrock model handling
